requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
instagrapi==1.19.1
multidict==6.0.4
openai==0.27.8
Pillow==10.0.0
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
requests==2.31.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
